requirements.txt
bitarray==2.7.3
imutils==0.5.4
numpy==1.24.2
opencv-python==4.7.0.72
pdf2image==1.16.3
pdfreader==0.1.12
Pillow==9.4.0
pycryptodome==3.17
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.8.2
pyzbar==0.1.9
qrcode==7.4.2
reportlab==3.6.12
six==1.16.0
typing_extensions==4.5.0
